+++ src/main/java/com/example/team10searchengine/configuration/batch/CsvWriter.java
@@ -1,7 +1,7 @@
-package com.example.team10searchengine.configuration;
+package com.example.team10searchengine.configuration.batch;
 
-import com.example.team10searchengine.entity.kordict.KorDict;
-import com.example.team10searchengine.entity.kordict.KorDictRepository;
+import com.example.team10searchengine.kordict.entity.KorDict;
+import com.example.team10searchengine.kordict.repository.jpqlrepo.KorDictRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.item.ItemWriter;
@@ -16,7 +16,7 @@ import java.util.List;
 public class CsvWriter implements ItemWriter<KorDict> {
 
     private final KorDictRepository korDictRepository;
-//    private final WekeRepository wekeRepository;
+
 
     @Override
     public void write(List<? extends KorDict> list) {
