+++ src/main/java/com/example/team10searchengine/kordict/entity/KorDictPrimary.java
@@ -1,4 +1,4 @@
-package com.example.team10searchengine.entity.kordict;
+package com.example.team10searchengine.kordict.entity;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
